requirements.txt
flask>=3.0.0
gitpython>=3.1.40
google-generativeai>=0.7.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
